1. Modify to pass USB-IF CV - Chapter 9 & Class test. 2. QSPI/SPI/USPI driver: Add xxx_TRIGGER_TX_RX_PDMA and xxx_DISABLE_TX_RX_PDMA API. 3. Correct USBD_VCOM_DualPort Descriptor. 4. Fix page program function issue of SPI Flash related sample code.(missing the 256th byte write)

+++ SampleCode/CardReader/USBD_CCID/ccid.h
@@ -1,9 +1,7 @@
 /******************************************************************************
  * @file     ccid.h
- * @brief    M451 series CCID driver header file
+ * @brief    CCID driver header file
  * @version  V1.00
- * $Revision: 1 $
- * $Date: 16/06/16 6:17p $
  *
  * @note
  * Copyright (C) 2014~2015 Nuvoton Technology Corp. All rights reserved.
@@ -130,7 +128,7 @@ __INLINE static uint32_t make32(uint8_t *buf)
 
 __INLINE static uint16_t make16(uint8_t *buf)
 {
-    return ((uint16_t) buf[1] << 8) | ((uint16_t) buf[0]);
+    return (uint16_t)(((uint16_t) buf[1] << 8) | ((uint16_t) buf[0]));
 }
 /*-------------------------------------------------------------*/
 extern uint8_t UsbIntMessageBuffer[];
